fix(web): rename citationFootnote utils for macOS E2E (#359)
On case-insensitive volumes, citationFootnotes.ts collided with
CitationFootnotes.tsx and broke mock-mode Playwright login. Add
INTERN-359 evidence doc and drop arbitrary sleep in intern upload spec.
