+++ web/src/components/qa/citationFootnoteUtils.ts
